Commit my local copy with github (3)
