+++ MontyHall/nbproject/project.properties
@@ -54,7 +54,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=montyhall.MontyHall
+main.class=ru.evsychev.main.MontyHall
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
