--- MontyHall/src/Games/Games.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Games;
-
-/**
- *
- * @author sychev
- */
-public interface Games {
-    
-    public static final int GAMES_COUNT_DEFAULT = 3;
-    
-    public void Start();
-    
-}
