+++ MontyHall/src/ru/evsychev/games/Games.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package ru.evsychev.games;
+
+/**
+ *
+ * @author sychev
+ */
+public interface Games {
+    
+    public static final int DOORS_COUNT_DEFAULT = 3;
+    
+    public void Start();
+    
+}
